feat: controller cliente start

+++ src/main/java/com/apollo/main/model/Pedido.java
@@ -15,7 +15,7 @@ public class Pedido {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idPedido;
+    private Long idPedido;
 
     @Basic
     @Enumerated(EnumType.STRING)
